refactor: move section.conf.js into src/config
