+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-07T15:28:26.835Z\n"
-"PO-Revision-Date: 2019-02-07T15:28:26.835Z\n"
+"POT-Creation-Date: 2019-02-07T15:44:40.667Z\n"
+"PO-Revision-Date: 2019-02-07T15:44:40.667Z\n"
 
 msgid "Cancel"
 msgstr ""
