+++ src/components/__tests__/AppRouter.test.js
@@ -5,7 +5,7 @@ import { shallow } from 'enzyme'
 import AppRouter from '../AppRouter'
 import { Route, Switch } from 'react-router-dom'
 
-import { sections } from '../../pages/sections.conf'
+import { sections } from '../../conf../../config/sections.conf'
 
 const pageState = {}
