+++ src/pages/StandardReport.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import Table from '@dhis2/d2-ui-table'
 import '@dhis2/d2-ui-core/build/css/Table.css'
 import '@dhis2/d2-ui-core/build/css/Pagination.css'
-import manageError from './manageError.HOC'
+import manageError from '../utils/pageEnhancers/manageError.HOC'
 import Feedback from '../components/Feedback'
 import Headline from './standard-report/Headline'
 import SearchBox from './standard-report/SearchBox'
