+++ src/pages/home/Home.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom'
 
 import MenuElement from '../../components/MenuElement'
 
-import { sections } from '../sections.conf'
+import { sections } from '../../config/sections.conf'
 import styles from './Home.style'
 
 const Home = () => {
